+++ pkg/reconciler/instances/istio/actions/performer.go
@@ -4,12 +4,12 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio-configuration/clientset"
-	"github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio-configuration/reset/proxy"
+	"github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio/clientset"
+	"github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio/reset/proxy"
 	"path/filepath"
 
-	"github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio-configuration/istioctl"
-	istioConfig "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio-configuration/reset/config"
+	"github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio/istioctl"
+	istioConfig "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio/reset/config"
 	reconcilerKubeClient "github.com/kyma-incubator/reconciler/pkg/reconciler/kubernetes"
 	"github.com/kyma-incubator/reconciler/pkg/reconciler/kubernetes/kubeclient"
 	"github.com/kyma-incubator/reconciler/pkg/reconciler/workspace"
