+++ pkg/reconciler/instances/istio/reset/data/gatherer.go
