+++ pkg/reconciler/instances/istio/reset/pod/handler.go
