+++ pkg/reconciler/instances/istio/reset/pod/reset/mocks/action.go
