+++ pkg/reconciler/instances/loader.go
@@ -10,7 +10,7 @@ import (
 	//import required to register component reconciler 'connectivityproxy' in reconciler registry
 	_ "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/connectivityproxy"
 	//import required to register component reconciler 'istio' in reconciler registry
-	_ "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio-configuration"
+	_ "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio"
 	//import required to register component reconciler 'ory' in reconciler registry
 	_ "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/ory"
 	//import required to register component reconciler 'rafter' in reconciler registry
